12/3/2014 – 12/4/2013:
Removed unnecessary variables and old/unnecessary/redundant lines of code from demo3 and EventHandler. Started working on better documentation for both files.

Continued to work on documentation for demo3. Changed the way it prints the demo prints the data at the end. Data is now printed on one line and prints the stimuli used in an orderly manner.

(This may be a redundant commit, but I was worried I might have messed something up with GIT and didn't want to push a bad version to the master. Apologies!)

-mm