KelpyScreenObject
This update includes the new KelpyScreenObject, a subclass that contains a pygame screen object plus space for extra attributes.
So far the only two included are WIDTH and HEIGHT, but this will likely change soon to include more usefull information.

The demo files (1-4) have all been updates to reflect this new change, and the initializaion function has been modified to now return a KelpyScreenObject rather than a pygame screen object.

-mm